Toggle week not finished

+++ rana/resources/schedule.py
@@ -10,39 +10,43 @@ schedule_blueprint = Blueprint('schedule', __name__)
 class ScheduleAPI(MethodView):
     def get(self, schedule_id):
         """Get the first week of the given schedule, or the week of the given monday date."""
-        schedule = Schedule.query.filter_by(id=schedule_id).first()
-        if schedule:
-            timeslots = None
-            if request.args.get('week'):
-                start_time = datetime.strptime(request.args.get('week'), '%Y-%m-%dT%H:%M:%S.%fZ')
-                end_time = start_time + timedelta(days=5)
-                timeslots = TimeSlot.query.with_parent(schedule).filter(TimeSlot.start_date.between(start_time, end_time)).all()
+        if schedule_id:
+            schedule = Schedule.query.filter_by(id=schedule_id).first()
+            if schedule:
+                timeslots = None
+                if request.args.get('week'):
+                    start_time = datetime.strptime(request.args.get('week'), '%Y-%m-%dT%H:%M:%S.%fZ')
+                    end_time = start_time + timedelta(days=5)
+                    timeslots = TimeSlot.query.with_parent(schedule).filter(TimeSlot.start_date.between(start_time, end_time)).all()
+                else:
+                    start_time = schedule.start_date
+                    end_time = start_time + timedelta(days=(5-start_time.weekday()))
+                    timeslots = TimeSlot.query.with_parent(schedule).filter(TimeSlot.start_date.between(start_time, end_time)).all()
+                resp = {
+                    'status': 'success',
+                    'name': schedule.name,
+                    'start_time': schedule.start_date.hour,
+                    'end_time': schedule.end_date.hour,
+                    'start_weekday': schedule.start_date.weekday(),
+                    'duration': schedule.duration,
+                    'timeslots': [{
+                            'id': ts.id, 
+                            'start_date': ts.start_date.strftime('%Y-%m-%dT%H:%M:%SZ'), 
+                            'duration': ts.duration, 
+                            'available': ts.available
+                        } for ts in timeslots]
+                }
+                return make_response(jsonify(resp)), 201
+                
             else:
-                start_time = schedule.start_date
-                end_time = start_time + timedelta(days=(5-start_time.weekday()))
-                timeslots = TimeSlot.query.with_parent(schedule).filter(TimeSlot.start_date.between(start_time, end_time)).all()
-            resp = {
-                'status': 'success',
-                'name': schedule.name,
-                'start_time': schedule.start_date.hour,
-                'end_time': schedule.end_date.hour,
-                'start_weekday': schedule.start_date.weekday(),
-                'duration': schedule.duration,
-                'timeslots': [{
-                        'id': ts.id, 
-                        'start_date': ts.start_date.strftime('%Y-%m-%dT%H:%M:%SZ'), 
-                        'duration': ts.duration, 
-                        'available': ts.available
-                    } for ts in timeslots]
-            }
-            return make_response(jsonify(resp)), 201
-            
+                resp = {
+                    'status': 'fail',
+                    'message': 'Schedule does not exist.',
+                }
+                return make_response(jsonify(resp)), 401
         else:
-            resp = {
-                'status': 'fail',
-                'message': 'Schedule does not exist.',
-            }
-            return make_response(jsonify(resp)), 401
+            # TODO: Implement report activity (sysadmin)
+            pass
 
     def post(self):
         """Create a new schedule."""
@@ -91,45 +95,49 @@ class ScheduleAPI(MethodView):
 
     def delete(self, schedule_id):
         """Delete a schedule by id, requires authorization."""
-        schedule = Schedule.query.filter_by(id=schedule_id).first()
-        if schedule:
-            sent_secret_code = None
-            if 'Authorization' in request.headers:
-                sent_secret_code = request.headers.get('Authorization')
-            else:
-                resp = {
-                    'status': 'fail',
-                    'message': 'Authorization failed. Please provide a secret code.'
-                }
-                return make_response(jsonify(resp)), 401
-            if sent_secret_code == schedule.secret_code:
-                try:
-                    TimeSlot.query.with_parent(schedule).delete(synchronize_session=False)
-                    db.session.delete(schedule)
-                    db.session.commit()
-                    resp = {
-                        'status': 'successs',
-                        'message': 'Schedule successfully deleted.'
-                    }
-                    return make_response(jsonify(resp)), 201
-                except Exception as e:
+        if schedule_id:
+            schedule = Schedule.query.filter_by(id=schedule_id).first()
+            if schedule:
+                sent_secret_code = None
+                if 'Authorization' in request.headers:
+                    sent_secret_code = request.headers.get('Authorization')
+                else:
                     resp = {
                         'status': 'fail',
-                        'message': e
+                        'message': 'Authorization failed. Please provide a secret code.'
+                    }
+                    return make_response(jsonify(resp)), 401
+                if sent_secret_code == schedule.secret_code:
+                    try:
+                        TimeSlot.query.with_parent(schedule).delete(synchronize_session=False)
+                        db.session.delete(schedule)
+                        db.session.commit()
+                        resp = {
+                            'status': 'successs',
+                            'message': 'Schedule successfully deleted.'
+                        }
+                        return make_response(jsonify(resp)), 201
+                    except Exception as e:
+                        resp = {
+                            'status': 'fail',
+                            'message': e
+                        }
+                        return make_response(jsonify(resp)), 401
+                else:
+                    resp = {
+                        'status': 'fail',
+                        'message': 'Authorization failed. Secret code is incorrect.'
                     }
                     return make_response(jsonify(resp)), 401
             else:
                 resp = {
                     'status': 'fail',
-                    'message': 'Authorization failed. Secret code is incorrect.'
+                    'message': 'Schedule does not exist.',
                 }
                 return make_response(jsonify(resp)), 401
         else:
-            resp = {
-                'status': 'fail',
-                'message': 'Schedule does not exist.',
-            }
-            return make_response(jsonify(resp)), 401
+            # TODO: Implement delete old schedules (sysadmin)
+            pass
 
 schedule_view = ScheduleAPI.as_view('schedule')
 
@@ -137,7 +145,7 @@ schedule_view = ScheduleAPI.as_view('schedule')
 schedule_blueprint.add_url_rule(
     '/schedule/',
     view_func=schedule_view,
-    methods=['POST',]
+    methods=['POST', 'DELETE', 'GET']
 )
 schedule_blueprint.add_url_rule(
     '/schedule/<string:schedule_id>',
